refactor(route-renderer): extract orchestrator prep services
Move page browser graph, global injector assets, and prepared-options assembly
into dedicated modules; route-render-orchestrator becomes a thin coordinator.
